# backend/openapi/openapi.public.yaml
openapi: 3.0.3
info:
  title: ChatXIV API (public)
  description: |
    Public HTTP surface for the ChatXIV backend (no admin routes).
    Update `backend/openapi/openapi.public.yaml` when you change public `/v1` contracts.

    Full API (including admin) is documented in `openapi.admin.yaml` and served at
    `GET /v1/admin/openapi.yaml` — see `/v1/admin/docs` in the browser
    (requires admin session cookie).
  version: 1.0.0

servers:
  - url: /
    description: Same host as this Swagger page (use for Try it out on dev/staging/prod — avoids calling localhost from the browser)
  - url: https://dev-api.chatxiv.com
    description: Development API host (when the spec is opened off-site or you want an absolute URL)
  - url: http://localhost:3000
    description: Local (browser address bar shows localhost)
  - url: http://127.0.0.1:3000
    description: Local numeric loopback (must match the page origin)

tags:
  - name: Health
  - name: Feature flags
  - name: Feedback
  - name: Documentation

paths:
  /health:
    get:
      tags: [Health]
      summary: Liveness probe
      operationId: getHealth
      responses:
        '200':
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok

  /v1/openapi.yaml:
    get:
      tags: [Documentation]
      summary: Public OpenAPI document (YAML)
      description: |
        Returns this public specification only. For the full API (admin routes), use
        `GET /v1/admin/openapi.yaml` (requires admin session).
      operationId: getOpenApiPublicYaml
      responses:
        '200':
          description: Public OpenAPI YAML document
          content:
            application/yaml:
              schema:
                type: string

  /v1/flags:
    get:
      tags: [Feature flags]
      summary: List feature flags
      description: |
        Returns flags sorted by name in stable offset pages (public). Defaults: `page=1`, `pageSize=50`, max `pageSize=500`.
      operationId: listFeatureFlags
      parameters:
        - $ref: '#/components/parameters/ListPage'
        - $ref: '#/components/parameters/ListPageSize'
      responses:
        '200':
          description: Paginated flag entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedFeatureFlagList'
        '400':
          description: Invalid pagination query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /v1/flags/{name}:
    get:
      tags: [Feature flags]
      summary: Get one feature flag
      description: |
        Returns the flag state. Unknown names return `{ name, enabled: false }` without `updatedAt`.
      operationId: getFeatureFlag
      parameters:
        - $ref: '#/components/parameters/FlagName'
      responses:
        '200':
          description: Flag entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagEntry'
        '400':
          description: Invalid flag name format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /v1/feedback:
    post:
      tags: [Feedback]
      summary: Submit feedback for a message
      description: |
        Accepts thumbs-up/down feedback with an idempotency key.
        Duplicate submissions (same key) return 200 without re-inserting.
      operationId: submitFeedback
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackBody'
      responses:
        '200':
          description: Feedback accepted (new or duplicate)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '400':
          description: Validation error (missing/invalid body or Idempotency-Key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated opaque string (e.g. UUID) for idempotent submission.
      schema:
        type: string
        example: 550e8400-e29b-41d4-a716-446655440000

    FlagName:
      name: name
      in: path
      required: true
      description: |
        ASCII letters (any case, including PascalCase), digits, and internal `.`, `_`, or `-`. Must start and end with a letter or digit.
      schema:
        type: string
        pattern: '^[a-zA-Z0-9](?:[a-zA-Z0-9._-]*[a-zA-Z0-9])?$'
        example: MyFeature

    ListPage:
      name: page
      in: query
      required: false
      description: 1-based page index.
      schema:
        type: integer
        minimum: 1
        default: 1

    ListPageSize:
      name: pageSize
      in: query
      required: false
      description: Page size (default 50; maximum 500).
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50

  schemas:
    FeatureFlagEntry:
      type: object
      required: [name, enabled]
      properties:
        name:
          type: string
        enabled:
          type: boolean
        updatedAt:
          type: string
          format: date-time
          description: Present when the flag exists in the store; omitted for unknown names on public GET.

    PaginatedFeatureFlagList:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FeatureFlagEntry'
        total:
          type: integer
          minimum: 0
          description: Total number of flags across all pages.
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1

    FeedbackBody:
      type: object
      required: [messageId, rating, reasonCode]
      properties:
        messageId:
          type: string
          description: Assistant message ID from the chat response.
        rating:
          type: string
          enum: [up, down]
        reasonCode:
          type: string
          enum: [wrong_answer, outdated, missing_info, unrelated, other]
          description: Required context when rating is "down".
        freeText:
          type: string
          maxLength: 500
          description: Optional free-text elaboration (max 500 chars).
        category:
          type: string
          description: |
            Optional chat context category; must be one of the `UsageCategory` string values
            from shared usage analytics types (e.g. RAIDING, MSQ), not an arbitrary string.

    FeedbackResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true

    ApiErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - VALIDATION_ERROR
            - RATE_LIMITED
            - REQUEST_TIMEOUT
            - PAYLOAD_TOO_LARGE
            - SOURCE_UNAVAILABLE
            - NO_SOURCED_ANSWER
            - REFUSAL
            - INTERNAL_ERROR
        message:
          type: string
        requestId:
          type: string
        stack:
          type: string
          description: Present only in non-production environments.